Refactor: homepage header recently played items
