+++ src/app/App.tsx
@@ -3,7 +3,7 @@ import { Route, Routes } from 'react-router-dom'
 import Sidebar from '../components/sidebar/Sidebar'
 import Content from '../components/Content'
 
-import Home from './pages/Home'
+import Home from './pages/home/Home'
 import Search from './pages/Search'
 import Account from './pages/Account'
 import Download from './pages/Download'
